fix: flow create project modal in accelerator
