+++ src/main/menu.ts
@@ -1,4 +1,5 @@
 /* eslint-disable no-console */
+import { useOpenPLCStore } from '@root/renderer/store'
 import { BrowserWindow, Menu, MenuItemConstructorOptions, nativeTheme, shell } from 'electron'
 
 import { i18n } from '../utils/i18n'
@@ -48,7 +49,10 @@ export default class MenuBuilder {
     return menu
   }
   handleCreateProject() {
-    this.mainWindow.webContents.send('project:create-accelerator')
+    const {
+      workspace: { editingState },
+    } = useOpenPLCStore()
+    this.mainWindow.webContents.send('project:create-accelerator', editingState) // Enviando o editingState para o main process
   }
 
   async handleOpenProject() {
